move `NginxService` into api as global module dependency
to make plumbing easier :(
