+++ api/src/unraid-api/nginx/nginx.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { NginxService } from '@app/unraid-api/nginx/nginx.service.js';
+
+/**
+ *
+ */
+@Module({
+    providers: [NginxService],
+    exports: [NginxService],
+})
+export class NginxModule {}
