+++ api/src/unraid-api/plugin/global-deps.module.ts
@@ -5,6 +5,7 @@ import { GRAPHQL_PUBSUB_TOKEN } from '@unraid/shared/pubsub/graphql.pubsub.js';
 import {
     API_KEY_SERVICE_TOKEN,
     LIFECYCLE_SERVICE_TOKEN,
+    NGINX_SERVICE_TOKEN,
     UPNP_CLIENT_TOKEN,
 } from '@unraid/shared/tokens.js';
 
@@ -12,12 +13,14 @@ import { pubsub } from '@app/core/pubsub.js';
 import { LifecycleService } from '@app/unraid-api/app/lifecycle.service.js';
 import { ApiKeyService } from '@app/unraid-api/auth/api-key.service.js';
 import { ApiKeyModule } from '@app/unraid-api/graph/resolvers/api-key/api-key.module.js';
+import { NginxModule } from '@app/unraid-api/nginx/nginx.module.js';
+import { NginxService } from '@app/unraid-api/nginx/nginx.service.js';
 import { upnpClient } from '@app/upnp/helpers.js';
 
 // This is the actual module that provides the global dependencies
 @Global()
 @Module({
-    imports: [ApiKeyModule],
+    imports: [ApiKeyModule, NginxModule],
     providers: [
         {
             provide: UPNP_CLIENT_TOKEN,
@@ -31,6 +34,10 @@ import { upnpClient } from '@app/upnp/helpers.js';
             provide: API_KEY_SERVICE_TOKEN,
             useClass: ApiKeyService,
         },
+        {
+            provide: NGINX_SERVICE_TOKEN,
+            useClass: NginxService,
+        },
         PrefixedID,
         LifecycleService,
         {
@@ -42,6 +49,7 @@ import { upnpClient } from '@app/upnp/helpers.js';
         UPNP_CLIENT_TOKEN,
         GRAPHQL_PUBSUB_TOKEN,
         API_KEY_SERVICE_TOKEN,
+        NGINX_SERVICE_TOKEN,
         PrefixedID,
         LIFECYCLE_SERVICE_TOKEN,
         LifecycleService,
