+++ packages/unraid-shared/src/services/nginx.ts
@@ -1,0 +1,7 @@
+export interface NginxService {
+    /**
+     * Reloads nginx via its rc script
+     * @returns true if the reload was successful, false otherwise
+     */
+    reload(): Promise<boolean>;
+}
