+++ packages/unraid-shared/src/tokens.ts
@@ -1,3 +1,4 @@
 export const UPNP_CLIENT_TOKEN = 'UPNP_CLIENT'; 
 export const API_KEY_SERVICE_TOKEN = 'ApiKeyService';
 export const LIFECYCLE_SERVICE_TOKEN = 'LifecycleService';
+export const NGINX_SERVICE_TOKEN = 'NginxService';
